+++ automaton/ColonDashAutomaton.h
@@ -6,10 +6,10 @@
 
 class ColonDashAutomaton : public Automaton {
 private:
-    static AutomatonResult s1(const string& input, int initialIndex, int currIndex, int currLine);
+    static AutomatonResult s1(const string& input, int initialIndex, size_t currIndex, int currLine);
 
 public:
-    AutomatonResult s0(const string& input, int currIndex, int currLine) override;
+    AutomatonResult s0(const string& input, size_t currIndex, int currLine) override;
     TestResult testAutomaton() override;
 };
